# pkg/bss-workplace-api/bss_workplace_api.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: ozonmp/bss_workplace_api/v1/bss_workplace_api.proto

/*
Package bss_workplace_api is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package bss_workplace_api

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_BssWorkplaceApiService_CreateWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, client BssWorkplaceApiServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CreateWorkplaceV1Request
	var metadata runtime.ServerMetadata

	newReader, berr := utilities.IOReaderFactory(req.Body)
	if berr != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
	}
	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq.Foo); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.CreateWorkplaceV1(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_BssWorkplaceApiService_CreateWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, server BssWorkplaceApiServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq CreateWorkplaceV1Request
	var metadata runtime.ServerMetadata

	newReader, berr := utilities.IOReaderFactory(req.Body)
	if berr != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
	}
	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq.Foo); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.CreateWorkplaceV1(ctx, &protoReq)
	return msg, metadata, err

}

func request_BssWorkplaceApiService_DescribeWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, client BssWorkplaceApiServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq DescribeWorkplaceV1Request
	var metadata runtime.ServerMetadata

	var (
		val string
		ok  bool
		err error
		_   = err
	)

	val, ok = pathParams["workplace_id"]
	if !ok {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "workplace_id")
	}

	protoReq.WorkplaceId, err = runtime.Uint64(val)
	if err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "workplace_id", err)
	}

	msg, err := client.DescribeWorkplaceV1(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_BssWorkplaceApiService_DescribeWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, server BssWorkplaceApiServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq DescribeWorkplaceV1Request
	var metadata runtime.ServerMetadata

	var (
		val string
		ok  bool
		err error
		_   = err
	)

	val, ok = pathParams["workplace_id"]
	if !ok {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "workplace_id")
	}

	protoReq.WorkplaceId, err = runtime.Uint64(val)
	if err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "workplace_id", err)
	}

	msg, err := server.DescribeWorkplaceV1(ctx, &protoReq)
	return msg, metadata, err

}

func request_BssWorkplaceApiService_ListWorkplacesV1_0(ctx context.Context, marshaler runtime.Marshaler, client BssWorkplaceApiServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ListWorkplacesV1Request
	var metadata runtime.ServerMetadata

	msg, err := client.ListWorkplacesV1(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_BssWorkplaceApiService_ListWorkplacesV1_0(ctx context.Context, marshaler runtime.Marshaler, server BssWorkplaceApiServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ListWorkplacesV1Request
	var metadata runtime.ServerMetadata

	msg, err := server.ListWorkplacesV1(ctx, &protoReq)
	return msg, metadata, err

}

func request_BssWorkplaceApiService_RemoveWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, client BssWorkplaceApiServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq RemoveWorkplaceV1Request
	var metadata runtime.ServerMetadata

	var (
		val string
		ok  bool
		err error
		_   = err
	)

	val, ok = pathParams["workplace_id"]
	if !ok {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "workplace_id")
	}

	protoReq.WorkplaceId, err = runtime.Uint64(val)
	if err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "workplace_id", err)
	}

	msg, err := client.RemoveWorkplaceV1(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_BssWorkplaceApiService_RemoveWorkplaceV1_0(ctx context.Context, marshaler runtime.Marshaler, server BssWorkplaceApiServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq RemoveWorkplaceV1Request
	var metadata runtime.ServerMetadata

	var (
		val string
		ok  bool
		err error
		_   = err
	)

	val, ok = pathParams["workplace_id"]
	if !ok {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "workplace_id")
	}

	protoReq.WorkplaceId, err = runtime.Uint64(val)
	if err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "workplace_id", err)
	}

	msg, err := server.RemoveWorkplaceV1(ctx, &protoReq)
	return msg, metadata, err

}

// RegisterBssWorkplaceApiServiceHandlerServer registers the http handlers for service BssWorkplaceApiService to "mux".
// UnaryRPC     :call BssWorkplaceApiServiceServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterBssWorkplaceApiServiceHandlerFromEndpoint instead.
func RegisterBssWorkplaceApiServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server BssWorkplaceApiServiceServer) error {

	mux.Handle("POST", pattern_BssWorkplaceApiService_CreateWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/CreateWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_BssWorkplaceApiService_CreateWorkplaceV1_0(rctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_CreateWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_BssWorkplaceApiService_DescribeWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/DescribeWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces/{workplace_id}"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_BssWorkplaceApiService_DescribeWorkplaceV1_0(rctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_DescribeWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_BssWorkplaceApiService_ListWorkplacesV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/ListWorkplacesV1", runtime.WithHTTPPathPattern("/v1/workplaces"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_BssWorkplaceApiService_ListWorkplacesV1_0(rctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_ListWorkplacesV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("DELETE", pattern_BssWorkplaceApiService_RemoveWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/RemoveWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces/{workplace_id}"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_BssWorkplaceApiService_RemoveWorkplaceV1_0(rctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_RemoveWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

// RegisterBssWorkplaceApiServiceHandlerFromEndpoint is same as RegisterBssWorkplaceApiServiceHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterBssWorkplaceApiServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.Dial(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterBssWorkplaceApiServiceHandler(ctx, mux, conn)
}

// RegisterBssWorkplaceApiServiceHandler registers the http handlers for service BssWorkplaceApiService to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterBssWorkplaceApiServiceHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterBssWorkplaceApiServiceHandlerClient(ctx, mux, NewBssWorkplaceApiServiceClient(conn))
}

// RegisterBssWorkplaceApiServiceHandlerClient registers the http handlers for service BssWorkplaceApiService
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "BssWorkplaceApiServiceClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "BssWorkplaceApiServiceClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "BssWorkplaceApiServiceClient" to call the correct interceptors.
func RegisterBssWorkplaceApiServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client BssWorkplaceApiServiceClient) error {

	mux.Handle("POST", pattern_BssWorkplaceApiService_CreateWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/CreateWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_BssWorkplaceApiService_CreateWorkplaceV1_0(rctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_CreateWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_BssWorkplaceApiService_DescribeWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/DescribeWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces/{workplace_id}"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_BssWorkplaceApiService_DescribeWorkplaceV1_0(rctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_DescribeWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_BssWorkplaceApiService_ListWorkplacesV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/ListWorkplacesV1", runtime.WithHTTPPathPattern("/v1/workplaces"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_BssWorkplaceApiService_ListWorkplacesV1_0(rctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_ListWorkplacesV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("DELETE", pattern_BssWorkplaceApiService_RemoveWorkplaceV1_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/ozonmp.bss_workplace_api.v1.BssWorkplaceApiService/RemoveWorkplaceV1", runtime.WithHTTPPathPattern("/v1/workplaces/{workplace_id}"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_BssWorkplaceApiService_RemoveWorkplaceV1_0(rctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_BssWorkplaceApiService_RemoveWorkplaceV1_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_BssWorkplaceApiService_CreateWorkplaceV1_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "workplaces"}, ""))

	pattern_BssWorkplaceApiService_DescribeWorkplaceV1_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1", "workplaces", "workplace_id"}, ""))

	pattern_BssWorkplaceApiService_ListWorkplacesV1_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "workplaces"}, ""))

	pattern_BssWorkplaceApiService_RemoveWorkplaceV1_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2}, []string{"v1", "workplaces", "workplace_id"}, ""))
)

var (
	forward_BssWorkplaceApiService_CreateWorkplaceV1_0 = runtime.ForwardResponseMessage

	forward_BssWorkplaceApiService_DescribeWorkplaceV1_0 = runtime.ForwardResponseMessage

	forward_BssWorkplaceApiService_ListWorkplacesV1_0 = runtime.ForwardResponseMessage

	forward_BssWorkplaceApiService_RemoveWorkplaceV1_0 = runtime.ForwardResponseMessage
)